+++ stackscripts/linode/hns-dane-appliance-bootstrap.sh
@@ -12,7 +12,7 @@ exec > >(tee -a "$LOG_FILE") 2>&1
 
 export DEBIAN_FRONTEND=noninteractive
 
-APPLIANCE_VERSION="v0.1.8"
+APPLIANCE_VERSION="v0.2.0"
 APPLIANCE_ARCHIVE_URL="https://github.com/denuoweb/dane-record-generator/archive/refs/tags/${APPLIANCE_VERSION}.tar.gz"
 APPLIANCE_ARCHIVE_SHA256="REPLACE_WITH_RELEASE_TARBALL_SHA256"